\name{pkgKitten-package}
\alias{pkgKitten-package}
\alias{pkgKitten}
\docType{package}
\title{
  What the package does in one short line
}
\description{
  More about what it does in maybe more than one line.

  In other words, a concise description of the package
  covering around one to five lines.
}
\details{
  \tabular{ll}{
    Package: \tab pkgKitten\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2014-04-01\cr
    License: \tab What license is it under?\cr
  }
  An overview of how to use the package, including the most important functions 
}
\author{
  Who wrote it

  Who maintains it, if that is a different person.
}
\references{
  Literature or other references for background information.
}
% Optionally other standard keywords, one per line, from file KEYWORDS
% in the R documentation directory 
\keyword{ package }
\seealso{
  Optional links to other man pages
  % E.g. \code{\link[zoo:zoo-package]{zoo}}
  % but note that the package must then be present in DESCRIPTION and NAMESPACE
}
\examples{
  #simple examples of the most important functions
  \dontrun{
     # use for code to display but not run
  }
}
