% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgKitten.R
\name{playWithPerPackageHelpPage}
\alias{playWithPerPackageHelpPage}
\title{Create a very simple package help page}
\usage{
playWithPerPackageHelpPage(name = "anRpackage", path = ".",
  maintainer = "Your Name", email = "your@mail.com")
}
\arguments{
\item{name}{The name of the package to be created, defaults to \dQuote{anRpackage}}

\item{path}{The path to the location where the package is to be
created, defaults to the current directory.}

\item{maintainer}{The name of the maintainer, defaults to
\dQuote{Your Name} or \code{author} if the latter is given.}

\item{email}{The maintainer email address.}
}
\value{
Nothing is returned as the function is invoked for its
side effect of creating a new package.
}
\description{
The \code{playWithPerPackageHelpPage} function creates an basic
package help page.
}
\details{
The \code{playWithPerPackageHelpPage} function can be used to
create a simple help page for a package.

It has been split off from the \code{kitten} function so that it
can be called from other packages. As such, it is also exported
from \pkg{pkgKitten}.
}
\author{
Dirk Eddelbuettel
}
