\name{pkgKitten-package}
\alias{pkgKitten-package}
\alias{pkgKitten}
\docType{package}
\title{
  Create simple packages using the \code{kitten} function.
}
\description{
  The base R function \code{package.skeleton()} is very helpful, but
  insist on creating packages which do not pass \code{R CMD check}.

  That is silly,

  This package helps with this issue. No more, no less.
}
\author{
  Dirk Eddelbuettel wrote the package and maintains it.
}
\keyword{package}
\seealso{
  \code{\link[utils:package.skeleton]{package.skeleton}},
  the \sQuote{Writing R Extensions} manual.
}
